// src/generated/ipc/TaskCreationConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DestinationTemplates } from "./DestinationTemplates";
import type { UnprojectedDestination } from "./UnprojectedDestination";

export type TaskCreationConfig = {
  defaultDestination: UnprojectedDestination;
  inboxFile: string;
  journalFolder: string;
  dailyFilenamePattern: string;
  projectFolder: string;
  templates: DestinationTemplates;
  migrationSource: string | null;
};